requirements.txt
geopy
snap-stanford
networkx
tqdm

